// packages/core/src/payload-types.ts
export interface User {
  id: string;
  username?: string | null;
  avatarUrl?: string | null;
  onboarded?: boolean | null;
  role?: ('admin' | 'user' | 'demo')[] | null;
  tenants?:
    | {
        tenant: string | Tenant;
        role: string | Role;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tenants".
 */
export interface Tenant {
  id: string;
  name: string;
  slug: string;
  subdomain: string;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
// ...
export interface Project {
  id: string;
  _order?: string | null;
  tenant?: (string | null) | Tenant;
  /**
   * Enter the name of the project.
   */
  name: string;
  /**
   * Provide a brief description of the project.
   */
  description?: string | null;
  /**
   * Attach a server, all the servers in this project will be deployed in that server
   */
  server: string | Server;
  services?: {
    docs?: (string | Service)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  /**
   * Hide this project from the public view.
   */
  hidden?: boolean | null;
  createdBy?: (string | null) | User;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
// ...
export interface Service {
  id: string;
  tenant?: (string | null) | Tenant;
  /**
   * Select the project associated with this service.
   */
  project: string | Project;
  backup?: (string | null) | Backup;
  /**
   * Enter the name of the service.
   */
  name: string;
  /**
   * Provide a brief description of the service.
   */
  description?: string | null;
  type: 'database' | 'app' | 'docker';
  variables?:
    | {
        key: string;
        value?: string | null;
        id?: string | null;
      }[]
    | null;
  volumes?:
    | {
        hostPath: string;
        containerPath: string;
        created?: boolean | null;
        id?: string | null;
      }[]
    | null;
  populatedVariables?: string | null;
  builder?: ('buildPacks' | 'railpack' | 'nixpacks' | 'dockerfile' | 'herokuBuildPacks' | 'static') | null;
  provider?: (string | null) | GitProvider;
  providerType?: ('github' | 'gitlab' | 'bitbucket' | 'azureDevOps' | 'gitea') | null;
  githubSettings?: {
    repository: string;
    owner: string;
    gitToken?: string | null;
    branch: string;
    buildPath: string;
    port?: number | null;
  };
  azureSettings?: {
    repository: string;
    branch: string;
    gitToken: string;
    owner: string;
    buildPath: string;
    port?: number | null;
  };
  giteaSettings?: {
    repository: string;
    branch: string;
    gitToken?: string | null;
    owner: string;
    buildPath: string;
    port?: number | null;
  };
  gitlabSettings?: {
    repository: string;
    branch: string;
    gitToken?: string | null;
    owner: string;
    buildPath: string;
    port?: number | null;
  };
  bitbucketSettings?: {
    repository: string;
    owner: string;
    gitToken?: string | null;
    branch: string;
    buildPath: string;
    port?: number | null;
  };
  databaseDetails?: {
    type?: ('postgres' | 'mongo' | 'mysql' | 'mariadb' | 'redis' | 'clickhouse') | null;
    username?: string | null;
    password?: string | null;
    host?: string | null;
    port?: string | null;
    connectionUrl?: string | null;
    version?: string | null;
    status?: ('running' | 'missing' | 'exited') | null;
    exposedPorts?: string[] | null;
  };
  dockerDetails?: {
    /**
     * Enter the docker-registry URL: ghrc://contentql/pin-bolt:latest
     */
    url?: string | null;
    account?: (string | null) | DockerRegistry;
    ports?:
      | {
          hostPort: number;
          containerPort: number;
          scheme: 'http' | 'https';
          id?: string | null;
        }[]
      | null;
  };
  domains?:
    | {
        domain: string;
        default: boolean;
        synced: boolean;
        autoRegenerateSSL?: boolean | null;
        certificateType?: ('letsencrypt' | 'none') | null;
        id?: string | null;
      }[]
    | null;
  deployments?: {
    docs?: (string | Deployment)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  traefikConfiguration?: {
    docs?: (string | Traefik)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "backups".
 */
export interface Backup {
  id: string;
  tenant?: (string | null) | Tenant;
  /**
   * Adding the service for which backup is related to
   */
  service: string | Service;
  type?: ('external' | 'internal') | null;
  backupName?: string | null;
  status: 'in-progress' | 'failed' | 'success';
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "gitProviders".
 */
export interface GitProvider {
  id: string;
  tenant?: (string | null) | Tenant;
  type: 'github' | 'gitlab' | 'bitbucket';
  github?: {
    appName: string;
    appUrl: string;
    appId: number;
    clientId: string;
    clientSecret: string;
    installationId?: string | null;
    privateKey: string;
    webhookSecret: string;
    installationToken?: string | null;
    tokenExpiration?: string | null;
  };
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "dockerRegistries".
 */
export interface DockerRegistry {
  id: string;
  tenant?: (string | null) | Tenant;
  name: string;
  type: 'docker' | 'github' | 'digitalocean' | 'quay';
  username: string;
  password: string;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "deployments".
 */
export interface Deployment {
  id: string;
  /**
   * Adding the service for which deployment is related to
   */
  service: string | Service;
  status: 'queued' | 'building' | 'failed' | 'success';
  logs?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "traefik".
 */
export interface Traefik {
  id: string;
  /**
   * Add the service for which traefik-configuration relates to
   */
  service: string | Service;
  configuration:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}